# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Attendence Management System
  description: |-
    Attendence Management System is a time tracker and leave management app that lets you track work hours and leave approvals of all employees across projects.
  termsOfService: http://localhost:8090
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8080/api
...
paths:
  /v1/login:
    post:
      tags:
        - Authenticate
      summary: Login for super admin
      description: 'allows super admin to enter into AMS system using email and password'
      parameters:
        - name: authorization
          in: header
          description: Basic Authentication - Base64 encoded(userName:password)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid  value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /v1/authenticate:
    post:
      tags:
        - Authenticate
      description: ''
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                authCode:
                  type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - type: object
                            properties:
                              profile:
                                $ref: '#/components/schemas/Profile'
                          - $ref: '#/components/schemas/LoginResponse'
        '301':
          description: redirect to sign up page
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 301
                  response:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      redirectUri:
                        type: string
                        example: http://localhost:8080/signup
        '400':
          description: Invalid input parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  response:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      status:
                        type: number
                        example: 400
                      message:
                        type: string
                        example: invalid auth code
                      restStatus:
                        type: string
                        example: "BAD_REQUEST"
        '503':
          description: Microsoft Service Unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  response:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      status:
                        type: number
                        example: 400
                      message:
                        type: string
                        example: microsoft service unavailable
                      restStatus:
                        type: string
                        example: "SERVICE_UNAVAILABLE"
...
components:
  schemas:
...
    BadRequest:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid Credentials
        error:
          type: boolean
          example: false
...
    LoginResponse:
      type: object
      properties:
        profile:
          $ref: '#/components/schemas/Profile'
        expires_in:
          type: string
          example: 4416
        access_token:
          type: string
          example: eyJ0eXAiOiJKV1QiLCJub25jZSI6IjZOZnNhTXVBVW9iM2NSOWpHaUxGNU9uRzlVS09TZXM1TlAtLXJhOWE5SWsiLCJhbGciOiJSUzI1NiIsIng1dCI6Ii1LSTNROW5OUjdiUm9meG1lWm9YcWJIWkdldyIsImtpZCI6Ii1LSTNROW5OUjdiUm9meG1lWm9YcWJIWkdldyJ9
    ApiResponse:
      type: object
      properties:
        statuscode:
          type: integer
          format: int32
          example: 200
        response:
          type: boolean
          example: true
    Profile:
      type: object
      properties:
        employeeName:
          type: string
          example: Bharath Devarajan,
        employeeDesignation:
          type: string
          example: Associate Software Engineer ,
        employeeImage:
          type: string
        employeeEmail:
          type: string
          example: bharath.d@flyerssoft.com,
        employeeOfficialId:
          type: string
          example: FS001,
        employeeMobileNumber:
          type: string
          example: +919894820022
        employeeLocation:
          type: string
          example: Chennai,
        employeeId:
          type: number
          example: 1
...